backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
bcrypt==3.2.2
boto3==1.39.3
botocore==1.39.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
databases==0.9.0
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fashion-clip==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hub-sdk==0.0.20
huggingface-hub==0.33.1
idna==3.7
iniconfig==2.1.0
ipyplot==1.1.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.4.8
pybboxes==0.1.6
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
roboflow==1.2.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
sahi==0.11.28
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
triton==3.3.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.0.239
ultralytics-thop==2.0.14
ultralyticsplus==0.1.0
urllib3==2.5.0
uvicorn==0.34.2
uvloop==0.21.0
validators==0.35.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
